refactor(frontend): use isTokenEthereumNative where applicable
