+++ src/frontend/src/eth/utils/native-token.utils.ts
@@ -1,0 +1,3 @@
+import type { Token } from '$lib/types/token';
+
+export const isTokenEthereumNative = (token: Token): boolean => token.standard?.code === 'ethereum';
