+++ src/frontend/src/eth/utils/token.utils.ts
@@ -98,5 +98,3 @@ export const calculateUsdValues = ({
 		sumInUSD: amountInUSD + feeInUSD
 	};
 };
-
-export const isTokenEthereumNative = (token: Token): boolean => token.standard?.code === 'ethereum';
